DEV: move _trim_mean to utils
