+++ mne/time_frequency/tests/test_psd.py
@@ -6,8 +6,7 @@ from nose.tools import assert_true, assert_equal
 
 from mne import pick_types, Epochs, read_events
 from mne.io import RawArray, read_raw_fif
-from mne.fixes import _trim_mean
-from mne.utils import slow_test, run_tests_if_main
+from mne.utils import slow_test, run_tests_if_main, _trim_mean
 from mne.time_frequency import psd_welch, psd_multitaper, psd_array_welch
 
 base_dir = op.join(op.dirname(__file__), '..', '..', 'io', 'tests', 'data')
