+++ mne/utils.py
@@ -40,7 +40,7 @@ from .externals.six.moves import urllib
 from .externals.six import string_types, StringIO, BytesIO, integer_types
 from .externals.decorator import decorator
 
-from .fixes import _get_args, _trim_mean
+from .fixes import _get_args
 
 logger = logging.getLogger('mne')  # one selection here used across mne-python
 logger.propagate = False  # don't propagate (in case of multiple imports)
@@ -2414,6 +2414,52 @@ def _get_fast_dot():
     except ImportError:
         fast_dot = np.dot
     return fast_dot
+
+
+# nan-safe scipy trim_mean (used in mne.time_frequency.psd_welch)
+def _trim_mean(a, proportiontocut, axis=0):
+    """
+    Return mean of array after trimming distribution from both tails.
+    If `proportiontocut` = 0.1, slices off 'leftmost' and 'rightmost' 10% of
+    scores. The input is sorted before slicing. Slices off less if proportion
+    results in a non-integer slice index (i.e., conservatively slices off
+    `proportiontocut` ).
+
+    Parameters
+    ----------
+    a : array_like
+        Input array
+    proportiontocut : float
+        Fraction to cut off of both tails of the distribution
+    axis : int or None, optional
+        Axis along which the trimmed means are computed. Default is 0.
+        If None, compute over the whole array `a`.
+
+    Returns
+    -------
+    trim_mean : ndarray
+        Mean of trimmed array.
+    """
+    a = np.asarray(a)
+
+    if a.size == 0:
+        return np.nan
+
+    if axis is None:
+        a = a.ravel()
+        axis = 0
+
+    nobs = a.shape[axis]
+    lowercut = int(proportiontocut * nobs)
+    uppercut = nobs - lowercut
+    if (lowercut > uppercut):
+        raise ValueError("Proportion too big.")
+
+    atmp = np.partition(a, (lowercut, uppercut - 1), axis)
+
+    sl = [slice(None)] * atmp.ndim
+    sl[axis] = slice(lowercut, uppercut)
+    return np.nanmean(atmp[sl], axis=axis)
 
 
 def _get_reduction(reduction, axis=-1):
